Binary population and algorithm as a stream.

+++ src/main/scala/info/rnowak/simplega/algorithm/GeneticAlgorithmResult.scala
@@ -1,3 +1,3 @@
 package info.rnowak.simplega.algorithm
 
-case class GeneticAlgorithmResult(totalIterations: Int)
+case class GeneticAlgorithmResult(totalIterations: Long)
